Add print component for mpd goal calculator
